Moving transformers to versions module closer to where they are actually used. See #16
